Created objects. Added .toString-override.

+++ src/Car.java
@@ -112,7 +112,7 @@ public abstract class Car implements IMovable {
     }
 
     private void setX(double x) {
-        this.x = (int) x;
+        this.x = x;
     }
 
     public double getY() {
@@ -120,7 +120,7 @@ public abstract class Car implements IMovable {
     }
 
     private void setY(double y) {
-        this.y = (int) y;
+        this.y = y;
     }
 
     /**
